% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_datasets.R
\name{tcga_available}
\alias{tcga_available}
\title{See available}
\usage{
tcga_available()
}
\value{
A filesystem
}
\description{
Opes TCGA MC3 mutation data
}
\examples{
# Pull All Coding Mutations in TCGA GBM cohort
load_tcga_mutations() |>
  dplyr::filter(Disease == "GBM", Silent = FALSE) |>

}
