% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_datasets.R
\name{tcga_load_cohort_mutations}
\alias{tcga_load_cohort_mutations}
\title{Load all cohort into memory}
\usage{
tcga_load_cohort_mutations(study)
}
\arguments{
\item{study}{name of the study to load data from (or keyword 'everything' to get all the data)}
}
\value{
data.frame
}
\description{
Loads all TCGA MC3 mutation data for a given cohort
}
\examples{
# Pull All Coding Mutations in TCGA GBM cohort
tcga_load_cohort_mutations('GBM')

}
