% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_datasets.R
\name{tcga_open_mutation_datasets}
\alias{tcga_open_mutation_datasets}
\title{Open arrow connection to TCGA mutation data}
\usage{
tcga_open_mutation_datasets()
}
\value{
A FileSystemDataset object (can work with it like its any other database connection)
}
\description{
Opens TCGA MC3 mutation data
}
\examples{
# Pull All Coding Mutations in TCGA GBM cohort
tcga_open_mutation_datasets() |>
  dplyr::filter(Disease == "GBM", Silent = FALSE) |>
  dplyr::collect()

}
